Refactored så product ligger i de delte filer
