+++ src/main/java/dk/sdu/se_f22/productmodule/management/ProductAttribute.java
@@ -17,7 +17,7 @@ public enum ProductAttribute {
     SIZE("size"),
     CLOCKSPEED("clockSpeed");
 
-    String alias;
+    public String alias;
 
 
     ProductAttribute(String alias){
